Refactor internal functions
