+++ src/main/kotlin/KtLodestone.kt
@@ -11,20 +11,10 @@ import cloud.drakon.ktlodestone.profile.ClassJob
 import cloud.drakon.ktlodestone.profile.GearSet
 import cloud.drakon.ktlodestone.profile.Minions
 import cloud.drakon.ktlodestone.profile.Mounts
-import io.ktor.client.HttpClient
-import io.ktor.client.call.body
-import io.ktor.client.engine.java.Java
-import io.ktor.client.request.get
-import io.ktor.client.request.header
-import io.ktor.http.HttpHeaders
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.future.future
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.jsonObject
-import kotlinx.serialization.json.jsonPrimitive
-import org.jsoup.Jsoup
 
 object KtLodestone {
     /**
@@ -147,57 +137,4 @@ object KtLodestone {
      * @throws LodestoneException Thrown when The Lodestone returns an unknown error.
      */
     @JvmStatic fun getMountsAsync(id: Int) = GlobalScope.future { Mounts.getMounts(id) }
-
-    private val meta = Json.parseToJsonElement(
-        object {}::class.java.classLoader.getResource("lodestone-css-selectors/meta.json") !!
-            .readText()
-    )
-
-    private val userAgentDesktop =
-        meta.jsonObject["userAgentDesktop"] !!.jsonPrimitive.content
-    private val userAgentMobile =
-        meta.jsonObject["userAgentMobile"] !!.jsonPrimitive.content
-
-    private val ktorClient = HttpClient(Java)
-
-    internal suspend fun getLodestoneProfile(
-        id: Int,
-        endpoint: String? = null,
-        mobileUserAgent: Boolean = false,
-    ) = coroutineScope {
-        val url = if (endpoint == null) {
-            "https://eu.finalfantasyxiv.com/lodestone/character/${id}/"
-        } else {
-            "https://eu.finalfantasyxiv.com/lodestone/character/${id}/${endpoint}"
-        }
-
-        val request = ktorClient.get(url) {
-            header(
-                HttpHeaders.UserAgent, if (! mobileUserAgent) {
-                    userAgentDesktop
-                } else {
-                    userAgentMobile
-                }
-            )
-        }
-
-        return@coroutineScope when (request.status.value) {
-            200 -> Jsoup.parse(request.body() as String)
-            404 -> throw CharacterNotFoundException("A character with ID `${id}` could not be found on The Lodestone.")
-            else -> throw LodestoneException()
-        }
-    }
-
-    internal suspend fun getLodestoneProfilePaginated(endpoint: String) =
-        coroutineScope {
-            val request = ktorClient.get(endpoint) {
-                header(HttpHeaders.UserAgent, userAgentDesktop)
-            }
-
-            return@coroutineScope if (request.status.value == 200) {
-                Jsoup.parse(request.body() as String)
-            } else {
-                throw LodestoneException()
-            }
-        }
 }
